基础代码

+++ zzyl-common/src/main/java/com/zzyl/common/core/page/TableDataInfo.java
@@ -37,7 +37,7 @@ public class TableDataInfo<T> implements Serializable
      * @param list 列表数据
      * @param total 总记录数
      */
-    public TableDataInfo(List<T> list, long total)
+    public TableDataInfo(List<T> list, int total)
     {
         this.rows = list;
         this.total = total;
